Add functions for adding a result/output collector to the end of a BuildSpec.
We will need to implement a container/step that will run at the end of every build to collect and
validate the results and resources a TaskRun produces against the ones specified by the Task.

The AddOutputResources function builds up the Build step definition with the information required to
do this collection and verification.